Updated headers for new classes
